requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comtypes==1.1.14
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
magic-filter==1.0.9
MarkupSafe==2.1.2
MouseInfo==0.1.3
multidict==6.0.4
Pillow==9.4.0
psutil==5.9.4
PyAutoGUI==0.9.53
pycaw==20230322
pygame==2.3.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python3-xlib==0.15
pytweening==1.0.4
pytz==2023.3
pywhatkit==5.4
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
Werkzeug==2.2.3
wikipedia==1.4.0
yarl==1.8.2
